Add shared stage texture GUI to billboards
