+++ demoSrc/demo_SharedStagePlaneMesh.js
@@ -1,5 +1,5 @@
 import { SharedStagePlaneMesh, SharedStageTexture } from "../esm/index.js";
-import { initSceneSet } from "./common.js";
+import { initSceneSet, initSharedStageTextureGUI } from "./common.js";
 import { Text, Sprite } from "pixi.js";
 import GUI from "lil-gui";
 import { MeshBasicMaterial } from "three";
@@ -49,8 +49,8 @@ const initBillBoards = (scene) => {
   const canvasBoard2 = initPlane(scene, -30, material);
 
   const gui = new GUI();
-  initGUI(gui, canvasBoard, "board 01");
-  initGUI(gui, canvasBoard2, "board 02");
+  initSharedStageTextureGUI(gui, canvasBoard, "board 01");
+  initSharedStageTextureGUI(gui, canvasBoard2, "board 02");
 };
 
 const initPlane = (scene, positionX, material) => {
@@ -65,18 +65,6 @@ const initPlane = (scene, positionX, material) => {
   canvasBoard.cameraChaser.isLookingCameraHorizontal = true;
   scene.add(canvasBoard);
   return canvasBoard;
-};
-
-const initGUI = (gui, billboard, name) => {
-  const target = billboard.cloneTextureArea();
-  const onUpdate = () => {
-    billboard.updateTextureAreaAndUV(target);
-  };
-  const folder = gui.addFolder(name);
-  folder.add(target, "x", 0, 1024, 1).onChange(onUpdate);
-  folder.add(target, "y", 0, 1024, 1).onChange(onUpdate);
-  folder.add(target, "width", 0, 1024, 1).onChange(onUpdate);
-  folder.add(target, "height", 0, 1024, 1).onChange(onUpdate);
 };
 
 window.onload = onDomContentsLoaded;
